Move workspace_util.ts under utils/ directory
