+++ src/utils/workspace.ts
@@ -1,6 +1,6 @@
 import * as fs from 'fs';
 import * as path from 'path';
-import { Uri, WorkspaceFolder } from 'vscode';
+import { Uri, workspace, WorkspaceFolder } from 'vscode';
 
 // searches up the folder structure until it finds a Cargo.toml
 export function nearestParentWorkspace(
@@ -43,4 +43,46 @@ export function nearestParentWorkspace(
   }
 
   return curWorkspace;
+}
+
+// This is an intermediate, lazy cache used by `getOuterMostWorkspaceFolder`
+// and cleared when VSCode workspaces change.
+let _sortedWorkspaceFolders: string[] | undefined;
+
+function sortedWorkspaceFolders(): string[] {
+  if (!_sortedWorkspaceFolders && workspace.workspaceFolders) {
+    _sortedWorkspaceFolders = workspace.workspaceFolders
+      .map(folder => {
+        let result = folder.uri.toString();
+        if (result.charAt(result.length - 1) !== '/') {
+          result = result + '/';
+        }
+        return result;
+      })
+      .sort((a, b) => {
+        return a.length - b.length;
+      });
+  }
+  return _sortedWorkspaceFolders || [];
+}
+
+export function getOuterMostWorkspaceFolder(
+  folder: WorkspaceFolder,
+  options?: { cached: boolean },
+): WorkspaceFolder {
+  if (!options || !options.cached) {
+    _sortedWorkspaceFolders = undefined;
+  }
+
+  const sorted = sortedWorkspaceFolders();
+  for (const element of sorted) {
+    let uri = folder.uri.toString();
+    if (uri.charAt(uri.length - 1) !== '/') {
+      uri = uri + '/';
+    }
+    if (uri.startsWith(element)) {
+      return workspace.getWorkspaceFolder(Uri.parse(element)) || folder;
+    }
+  }
+  return folder;
 }
